feat: handle clock drift #14
Added drift tolerance of 10 seconds to preserve monotonicity when the
system clock is adjusted by NTP after a small clock drift or when the
system clock jumps back by 1 second due to leap second.

The counter is incremented when the current time:
- is the same as the previous time;
- has moved backwards up to 10 seconds.

The timestamp is also incremented if the counter is exceeded. Now the
TSID factory doesn't have to wait for the next millisecond when the
counter overruns. Therefore, there is no longer any speed limit for
generating TSIDs.